refactoring functions files
